backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
python-dotenv==1.0.0
Pillow==10.2.0
whitenoise==6.7.0
gunicorn==23.0.0
dj-database-url==2.3.0
psycopg2-binary==2.9.9
